# Conflicts:
#	features/user/chat_common/src/main/java/com/tokopedia/chat_common/domain/pojo/ChatItemPojo.kt
#	features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/fragment/TopChatRoomFragment.kt

+++ features/transaction/buyerorder/src/main/java/com/tokopedia/buyerorder/detail/view/fragment/MarketPlaceDetailFragment.java
@@ -1322,7 +1322,7 @@ public class MarketPlaceDetailFragment extends BaseDaggerFragment implements Ref
         for (Items item : items) {
             JsonObject passenger = new JsonObject();
 
-            int productId = 0;
+            long productId = 0;
             int quantity = 0;
             int shopId = 0;
             String notes = "";
